refactor: standardize serverUrl naming across multiple modules and remove unused SchnorrKeysign and DKLSKeysign classes for cleaner implementation
