+++ core/mpc/dkls/dkls.ts
@@ -81,7 +81,7 @@ export class DKLS {
         message?.receivers.forEach(receiver => {
           // send message to receiver
           sendRelayMessage({
-            serverURL: this.serverURL,
+            serverUrl: this.serverURL,
             localPartyId: this.localPartyId,
             sessionId: this.sessionId,
             message: messageToSend,
@@ -102,7 +102,7 @@ export class DKLS {
     while (true) {
       try {
         const downloadMsg = await downloadRelayMessage({
-          serverURL: this.serverURL,
+          serverUrl: this.serverURL,
           localPartyId: this.localPartyId,
           sessionId: this.sessionId,
         })
@@ -133,7 +133,7 @@ export class DKLS {
           }
           this.cache[cacheKey] = ''
           await deleteRelayMessage({
-            serverURL: this.serverURL,
+            serverUrl: this.serverURL,
             localPartyId: this.localPartyId,
             sessionId: this.sessionId,
             messageHash: msg.hash,
